fix many things
Fix session
Fix Index in table
Added add and remove in cart (Not working)
Added conditionnals in html with thymeleaf

+++ src/main/java/ch/heigvd/sprint0/repository/CartRepository.java
@@ -3,5 +3,5 @@ package ch.heigvd.sprint0.repository;
 import ch.heigvd.sprint0.model.Cart;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CartRepository extends CrudRepository<Cart, Long> {
+public interface CartRepository extends CrudRepository<Cart, Integer> {
 }
