+++ src/main/java/ch/heigvd/sprint0/service/ArticleService.java
@@ -30,12 +30,12 @@ public class ArticleService implements IArticleService {
         return (List<Article>) articleRepository.findAll();
     }
 
-    public Optional<Article> findById(long id) {
+    public Optional<Article> findById(int id) {
         return articleRepository.findById(id);
     }
 
     @Override
-    public List<Cart_Article> findCartArticleFromUser(long id) {
+    public List<Cart_Article> findCartArticleFromUser(int id) {
         return new LinkedList<>(cartRepository.findById(id).get().getCart_article_list());
     }
 }
